requirements.txt
asgiref==3.5.0
Django==3.2.12
django-appconf==1.0.5
django-imagekit==4.1.0
pilkit==2.0
Pillow==9.5.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
